updated README with changes to mapper
If a depth image uses 0 for an out-of-range value, then the mapper thinks there is an obstacle 0 meters away from the camera, which leaves a trail of obstacle voxels in the path of the camera. To circumvent this, add the line of code so all points with a depth of 0 are not added to the point cloud. This should not cause any errors with other choices of out-of-range values because cameras can't detect an object 0 meters away anyway.