Listens to tf for pose estimate instead of using AMCL
If you don't want to use AMCL's pose estimate which requires a prior map,
then the convert script will listen to a specified transform to produce a
new pose message. This is required if the AMCL-produced tf between /map
and /odom frames is overwritten with a static transform because current
pose topic is AMCL-dependent.